+++ R/itemUsageConstraint.R
@@ -11,11 +11,16 @@
 #' each item is used exactly once, which corresponds to no item-overlap and
 #' complete item pool depletion.
 #'
+#' If certain items are required in the resulting test form(s), as for example anchor items,
+#' \code{whichItems} can be used to constrain the usage of these items to be exactly 1.
+#' \code{whichItems} can either be a numeric vector with item numbers or a character vector
+#' with item identifiers corresponding to \code{itemIDs}.
+#'
 #'@inheritParams itemValuesConstraint
 #'@param nItems Number of items in the item pool [optional to create \code{itemIDs} automatically].
 #'@param formValues vector with values or weights for each form. Defaults to 1 for each form.
 #'@param targetValue The value to be used in the constraints
-#'@param whichItems A vector indicating which items should be constrained. Defaults to all the items forms.
+#'@param whichItems A vector indicating which items should be constrained. Defaults to all the items.
 #'
 #'@return An object of class \code{"constraint"}.
 #'
@@ -24,6 +29,13 @@
 #' ##  for 2 test forms with an item pool of 20 items
 #' itemUsageConstraint(2, operator = "=", targetValue = 1,
 #'                     itemIDs = 1:20)
+#'
+#' ## force certain items to be in the test, others not
+#' usage1 <- itemUsageConstraint(2, operator = "<=", targetValue = 1,
+#'                     itemIDs = paste0("item", 1:20))
+#' usage2 <- itemUsageConstraint(2, operator = "=", targetValue = 1,
+#'                     itemIDs = paste0("item", 1:20),
+#'                     whichItems = c("item5", "item8", "item10"))
 #'
 #'@export
 # constraints sum of item values over forms
@@ -47,6 +59,8 @@ itemUsageConstraint <- function(nForms, nItems = NULL, formValues = rep(1, nForm
   # the targetValue should be smaller than or equal to the sum of the formValues
   if(targetValue > sum(formValues)) stop("The 'targetValue' should be smaller than the sum of the 'formValues'.")
 
+  # item IDs supplied to whichItems
+  whichItems <- whichItemsToNumeric(whichItems, itemIDs = itemIDs)
 
   # choose info_text for info
   if(is.null(info_text)) info_text <- paste0(deparse(substitute(formValues)), operator, targetValue)
@@ -56,4 +70,15 @@ itemUsageConstraint <- function(nForms, nItems = NULL, formValues = rep(1, nForm
                      whichItems, sense = NULL,
                      info_text = info_text,
                      itemIDs = itemIDs)
+}
+
+whichItemsToNumeric <- function(whichItems, itemIDs) {
+  if(is.character(whichItems)) {
+    if(is.null(itemIDs)) stop("item IDs supplied to 'whichItems' but not to 'itemIDs'.")
+    not_in_itemIDs <- setdiff(whichItems, itemIDs)
+    if(length(not_in_itemIDs) > 0) stop("The following item IDs are in 'whichItem' but not in 'itemIDs': ",
+                                        paste(not_in_itemIDs, collapse = ", "))
+    return(match(whichItems, itemIDs))
+  }
+  whichItems
 }